# Disable Huffman encoding for small headers