--- a/codec-http2/src/test/java/io/netty/handler/codec/http2/DefaultHttp2FrameWriterTest.java
+++ b/codec-http2/src/test/java/io/netty/handler/codec/http2/DefaultHttp2FrameWriterTest.java
@@ -15,6 +15,7 @@
 package io.netty.handler.codec.http2;
 
 import io.netty.buffer.ByteBuf;
+import io.netty.buffer.ByteBufUtil;
 import io.netty.buffer.Unpooled;
 import io.netty.buffer.UnpooledByteBufAllocator;
 import io.netty.channel.Channel;
@@ -65,16 +66,17 @@ public class DefaultHttp2FrameWriterTest {
     @Before
     public void setUp() throws Exception {
         MockitoAnnotations.initMocks(this);
+        http2HeadersEncoder = new DefaultHttp2HeadersEncoder(
+                Http2HeadersEncoder.NEVER_SENSITIVE, new HpackEncoder(false, 16, 0));
 
-        frameWriter = new DefaultHttp2FrameWriter();
+        frameWriter = new DefaultHttp2FrameWriter(new DefaultHttp2HeadersEncoder(
+                Http2HeadersEncoder.NEVER_SENSITIVE, new HpackEncoder(false, 16, 0)));
 
         outbound = Unpooled.buffer();
 
         expectedOutbound = Unpooled.EMPTY_BUFFER;
 
         promise = new DefaultChannelPromise(channel, ImmediateEventExecutor.INSTANCE);
-
-        http2HeadersEncoder = new DefaultHttp2HeadersEncoder();
 
         Answer<Object> answer = new Answer<Object>() {
             @Override