@normanmaurer 2019-06-20
@netty-bot test this please