@carl-mastrangelo 2019-06-20
Also, some initial runs of the HpackDecoderBenchmark shows this is numerous times faster.