@normanmaurer 2019-06-20
This sounds like a good compromise 

> Am 20.06.2019 um 17:00 schrieb Eric Anderson <notifications@github.com>:
> 
> I agree most of the gain from HPACK is from reusing headers from the table (for things like user-agent and cookies), not huffman. However, the highest possible compression with huffman is a compression ratio of 5/8 = 63%, which is still useful.
> 
> How would y'all feel about choosing to do huffman only when the header size is > 50 or > 100 or > 1000, or some such? That would allow us to skip huffman for the majority of headers, but still do it when it can provide larger gains. Since HTTP requests don't tend to have lots of headers, it doesn't seem like there are enough small headers to add up to many bytes saved.
> 
> I will note that Huffman is also even more expensive than the benchmark shows, since the huffman cost is paid by both the sender and the receiver.
> 
> I do think it would be fair to have a constructor that disables huffman entirely, but I'd really only expect fully disabling huffman when communicating within a datacenter. When going over the internet or between continents, the CPU cost is probably worth it (for larger headers).
> 
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub, or mute the thread.