@carl-mastrangelo 2019-06-20
> How would y'all feel about choosing to do huffman only when the header size is > 50 or > 100 or > 1000, or some such?

I had considered this in an earlier revision, but didn't like it.   It means that as the headers get larger, the performance gets disproportionately worse.  Notice in the before / after numbers, there is a non linear speedup between the large and small headers.  

If I *had* to pick a number, I would say somewhere between 256 and 512.   384?

> I will note that Huffman is also even more expensive than the benchmark shows, since the huffman cost is paid by both the sender and the receiver.

Yes.  I ran the decoder benchmark briefly and saw about 6x speedup.