@normanmaurer 2019-06-21
@ejona86 @carl-mastrangelo maybe make it configurable and then use `384` as default ? Using `0` could have the meaning of disable all-together.