@normanmaurer 2019-06-21
@carl-mastrangelo also can you please attach a flame graph to understand why we burn all the CPU ? I assume it is because we basically call `writeByte(...)` in a loop which then will do all kind of reference checks.