@normanmaurer 2019-06-21
Also @Lukasa told me that there may be a better ( == faster ) way to implement this:

https://github.com/apple/swift-nio-http2/blob/master/Sources/NIOHPACK/HuffmanTables.swift

We should investigate if we should port this over @ejona86 @carl-mastrangelo