@Lukasa 2019-06-21
The Swift code is a good approximation to the implementation but probably obfuscates what it does. The more clear approach is to look at the Python implementation. The underlying data structure is defined [here](https://github.com/python-hyper/hpack/blob/master/hpack/huffman_table.py), and the code that uses the data structure is [here](https://github.com/python-hyper/hpack/blob/master/hpack/huffman.py).