@carl-mastrangelo 2019-06-21
> I assume it is because we basically call writeByte(...) in a loop which then will do all kind of reference checks.

There is a large amount of time here (I think the perfasm output is more illuminating), split between checking bounds each write, and checking the ByteBuf is accessible.

@Lukasa a table lookup would speed things up, I'll look through it.

[perfasm.txt](https://github.com/netty/netty/files/3315801/perfasm.txt)