@njhill 2019-06-21
@carl-mastrangelo @normanmaurer I had been meaning to raise this for a while, will open an issue to discuss more: performance of certain `ByteBuf`-accessing code is *highly* sensitive to stack depth, due to the default max inlining depth (of 9). This is not because of the inlining itself, but the resulting reordering/elision of accessibility checks once they have been inlined into the same method.

I'm fairly certain that's at play here... try running the same benchmark with e.g. `-XX:MaxInlineLevel=24`.

Unfortunately I can't think of a good solution to this apart from recommending users set the above JVM arg, since we have no control over where netty is called from within apps' code, which could already be deep. We could partially mitigate by trying to manually inline and shorten relevant call paths (I started to do this a bit within the `ByteBuf` class hierarchy), but doing so aggressively would harm code readability/duplication and only help the borderline cases anyhow.

I have seen comments about the default of 9 being outdated (unnecessarily small given other improvements), and wonder why this hasn't been increased in newer hotspot versions.