@normanmaurer 2019-06-22
@carl-mastrangelo @njhill what I expected :(  One idea I also had in the past was to introduce something like:

```
interface ByteTransformer {
    byte transform(byte b);
}

ByteBuf.transform(ByteTransformer, int, int)
```

This was we could do the range check / access check only once per transformation.

This of course only works for the case when we need to change stuff byte wise.