@normanmaurer 2019-06-25
@carl-mastrangelo do you work on the "table-lookup" change or should I try to find time over the next days ?