@carl-mastrangelo 2019-06-25
@normanmaurer TBH, I'm not that excited about making Huffman coding faster.  I'd rather make this PR (and perhaps leave the default to still do encoding), just to have an option to disable it entirely.