@normanmaurer 2019-06-25
@carl-mastrangelo I think there is not wrong with allow to disable it at least... So maybe lets just enable it by default but provide a way to disable it. We can do a followup to improve the encoding itself.