@normanmaurer 2019-06-27
@carl-mastrangelo I plan to cut the next release tomorrow so if you want to have this in this is the time ;)