@carl-mastrangelo 2019-06-27
@normanmaurer I have modified it to be a threshold, where strings above the threshold will be encoded, and below will not.   

It's probably okay if this misses a release.  If you are happy with the API and the implementation, I would be happy if this makes it into 4.1.37