@normanmaurer 2019-06-28
@carl-mastrangelo thanks... lets give @ejona86 a chance to review again first.