@carl-mastrangelo 2019-06-29
@normanmaurer Done!


The benchmarks are still correct, since the "LARGE" headers only go up to 300 byte values.