@carl-mastrangelo 2019-06-29
@normanmaurer & @Lukasa  For posterity, I filed https://github.com/netty/netty/pull/9307   to improve the decoder performance.